refactor: Move AgentStatus enum from service to data package to resolve circular import
