+++ src/ai/backend/manager/repositories/schedule/repository.py
@@ -32,6 +32,7 @@ from ai.backend.common.types import (
 )
 from ai.backend.logging.utils import BraceStyleAdapter
 from ai.backend.manager.config.provider import ManagerConfigProvider
+from ai.backend.manager.data.agent.types import AgentStatus
 from ai.backend.manager.data.kernel.types import KernelStatus
 from ai.backend.manager.data.session.types import SessionStatus
 from ai.backend.manager.decorators.repository_decorator import (
@@ -41,7 +42,6 @@ from ai.backend.manager.exceptions import ErrorStatusInfo
 from ai.backend.manager.models import (
     PRIVATE_SESSION_TYPES,
     AgentRow,
-    AgentStatus,
     DefaultForUnspecified,
     DomainRow,
     EndpointRow,
