+++ tests/manager/repositories/schedule/test_termination.py
@@ -27,12 +27,12 @@ from ai.backend.common.types import (
 )
 from ai.backend.manager.config.loader.legacy_etcd_loader import LegacyEtcdLoader
 from ai.backend.manager.config.provider import ManagerConfigProvider
+from ai.backend.manager.data.agent.types import AgentStatus
 from ai.backend.manager.data.kernel.types import KernelStatus
 from ai.backend.manager.data.session.types import SessionStatus
 from ai.backend.manager.data.user.types import UserRole
 from ai.backend.manager.models import (
     AgentRow,
-    AgentStatus,
     DomainRow,
     GroupRow,
     KernelRow,
